requirements.txt
PyGithub==1.43.7
github-stargazers==0.1.0
pandas==0.23.4
Google-Search-API==1.1.14
lxml
